requirements.txt
ebooklib
bs4
openai
python-dotenv
nltk
tiktoken